moved each view into its own component file
