+++ components/Views/AfterScanView.tsx
